+++ tests/e2e/single/vpc_test.go
@@ -3,14 +3,12 @@
 package single
 
 import (
-	"errors"
 	"fmt"
 	"strings"
 	"testing"
 	"time"
 
 	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/aws/aws-sdk-go/service/ec2"
 	"github.com/mulgadc/spinifex/tests/e2e/harness"
 	"github.com/stretchr/testify/require"
@@ -291,473 +289,4 @@ func runVPCSubnetE2E(t *testing.T, fix *Fixture) {
 		// see the WAN egress signal.
 		t.Logf("WARN: external ping inconclusive — may depend on WAN gateway")
 	}
-}
-
-// runRouteTableValidation ports run-e2e.sh Phase 8c (~2626-2745).
-//
-// Validates route-table CRUD + invariants against the *default* VPC's main
-// route table (always runs — no pool-mode gating). Three sub-tests:
-//
-//	Step1_DefaultMainRouteTable — DescribeRouteTables filter by vpc-id +
-//	    association.main asserts the local route is present. The IGW route
-//	    is verified only when the default VPC has an attached IGW (some
-//	    bootstraps skip it on private-only fixtures).
-//	Step2_CustomRouteTableLifecycle — CreateRouteTable, verify the implicit
-//	    local route, AssociateRouteTable to a default subnet, DisassociateRouteTable,
-//	    DeleteRoute on a manually-added route, DeleteRouteTable.
-//	Step3_ErrorPaths — DeleteRoute on the local route + DeleteRouteTable on
-//	    the main route table must both fail.
-//
-// Each step registers its own cleanup so a partial run still releases the
-// scratch route table / association regardless of which assertion fired.
-func runRouteTableValidation(t *testing.T, fix *Fixture) {
-	harness.Phase(t, "Single — Route Table Validation")
-
-	c := fix.AWS
-
-	// Default VPC + subnet + SG are owned by the harness Ensure* memo, so
-	// repeated calls across phases (and post-migration: across sibling
-	// tests) hit the same cached IDs without re-discovering.
-	def := harness.EnsureDefaultVPC(t, fix.Harness)
-	require.NotEmpty(t, def.VPCID, "default VPC ID required")
-	require.NotEmpty(t, def.SubnetID, "default subnet ID required")
-	harness.Detail(t, "vpc", def.VPCID, "subnet", def.SubnetID)
-
-	// --- Step 1: Default VPC main route table -----------------------------
-	var mainRTBID string
-	t.Run("Step1_DefaultMainRouteTable", func(t *testing.T) {
-		harness.Step(t, "describe-route-tables filter=vpc-id,association.main")
-		out, err := c.EC2.DescribeRouteTables(&ec2.DescribeRouteTablesInput{
-			Filters: []*ec2.Filter{
-				{Name: aws.String("vpc-id"), Values: []*string{aws.String(def.VPCID)}},
-				{Name: aws.String("association.main"), Values: []*string{aws.String("true")}},
-			},
-		})
-		require.NoError(t, err, "describe-route-tables (main)")
-		require.NotEmptyf(t, out.RouteTables,
-			"no main route table found for default VPC %s", def.VPCID)
-		mainRTBID = aws.StringValue(out.RouteTables[0].RouteTableId)
-		require.NotEmpty(t, mainRTBID, "main RouteTableId empty")
-		harness.Detail(t, "main_rtb", mainRTBID)
-
-		// GetRouteTableAssociations equivalent: confirm at least one
-		// association of the main RT is flagged Main=true. The aws-sdk-go
-		// surface folds this into RouteTable.Associations[].Main, so a
-		// separate API call isn't needed — but assert it explicitly so a
-		// regression that drops the Main flag would fail here, not silently.
-		mainFlagSeen := false
-		for _, a := range out.RouteTables[0].Associations {
-			if aws.BoolValue(a.Main) {
-				mainFlagSeen = true
-				break
-			}
-		}
-		require.Truef(t, mainFlagSeen,
-			"main route table %s has no Association with Main=true", mainRTBID)
-
-		// Local route is mandatory on every route table — the gateway
-		// auto-creates it at CreateRouteTable / CreateVpc time.
-		localFound := false
-		for _, r := range out.RouteTables[0].Routes {
-			if aws.StringValue(r.GatewayId) == "local" {
-				localFound = true
-				harness.Detail(t, "local_route", aws.StringValue(r.DestinationCidrBlock))
-				break
-			}
-		}
-		require.Truef(t, localFound,
-			"main route table %s missing local route", mainRTBID)
-
-		// IGW route (0.0.0.0/0 -> igw-...) — soft check. Default VPCs that
-		// admin init wires with an IGW must have it; private-only bootstraps
-		// legitimately don't, so just log when missing rather than fail.
-		igwRoute := ""
-		for _, r := range out.RouteTables[0].Routes {
-			if aws.StringValue(r.DestinationCidrBlock) == "0.0.0.0/0" {
-				igwRoute = aws.StringValue(r.GatewayId)
-				break
-			}
-		}
-		if igwRoute != "" {
-			harness.Detail(t, "default_route", "0.0.0.0/0->"+igwRoute)
-		} else {
-			t.Logf("note: default VPC main RT has no 0.0.0.0/0 route (no IGW attached)")
-		}
-	})
-
-	// Bail if Step 1 didn't seat mainRTBID — Steps 2/3 dereference it.
-	if mainRTBID == "" {
-		t.Fatalf("Step 1 did not resolve the main route table; cannot continue")
-	}
-
-	// --- Step 2: Custom route table CRUD lifecycle ------------------------
-	t.Run("Step2_CustomRouteTableLifecycle", func(t *testing.T) {
-		harness.Step(t, "create-route-table (vpc=%s)", def.VPCID)
-		ctOut, err := c.EC2.CreateRouteTable(&ec2.CreateRouteTableInput{
-			VpcId: aws.String(def.VPCID),
-		})
-		require.NoError(t, err, "create custom route table")
-		require.NotNil(t, ctOut.RouteTable, "create-route-table returned nil RouteTable")
-		customRTB := aws.StringValue(ctOut.RouteTable.RouteTableId)
-		require.NotEmpty(t, customRTB, "custom RouteTableId empty")
-		rtbDeleted := false
-		t.Cleanup(func() {
-			if !rtbDeleted {
-				_, _ = c.EC2.DeleteRouteTable(&ec2.DeleteRouteTableInput{
-					RouteTableId: aws.String(customRTB),
-				})
-			}
-		})
-		harness.Detail(t, "custom_rtb", customRTB)
-
-		// Local route must be present on the freshly-created table.
-		descCustom, err := c.EC2.DescribeRouteTables(&ec2.DescribeRouteTablesInput{
-			RouteTableIds: []*string{aws.String(customRTB)},
-		})
-		require.NoError(t, err, "describe custom route table")
-		require.NotEmpty(t, descCustom.RouteTables, "custom RT not visible after create")
-		customLocal := ""
-		for _, r := range descCustom.RouteTables[0].Routes {
-			if aws.StringValue(r.GatewayId) == "local" {
-				customLocal = aws.StringValue(r.DestinationCidrBlock)
-				break
-			}
-		}
-		require.NotEmptyf(t, customLocal,
-			"custom route table %s missing local route", customRTB)
-		harness.Detail(t, "custom_local_route", customLocal)
-
-		// Find an attached IGW for the default VPC so we can add a real
-		// 0.0.0.0/0 route. If none is attached the test still exercises
-		// associate/disassociate + delete on the local route only.
-		igwForVPC := ""
-		igws, err := c.EC2.DescribeInternetGateways(&ec2.DescribeInternetGatewaysInput{
-			Filters: []*ec2.Filter{
-				{Name: aws.String("attachment.vpc-id"), Values: []*string{aws.String(def.VPCID)}},
-			},
-		})
-		require.NoError(t, err, "describe-internet-gateways")
-		if len(igws.InternetGateways) > 0 {
-			igwForVPC = aws.StringValue(igws.InternetGateways[0].InternetGatewayId)
-			harness.Detail(t, "vpc_igw", igwForVPC)
-		}
-
-		routeAdded := false
-		if igwForVPC != "" {
-			harness.Step(t, "create-route 0.0.0.0/0 -> %s", igwForVPC)
-			_, err = c.EC2.CreateRoute(&ec2.CreateRouteInput{
-				RouteTableId:         aws.String(customRTB),
-				DestinationCidrBlock: aws.String("0.0.0.0/0"),
-				GatewayId:            aws.String(igwForVPC),
-			})
-			require.NoError(t, err, "create-route 0.0.0.0/0 -> IGW (custom RT)")
-			routeAdded = true
-			t.Cleanup(func() {
-				_, _ = c.EC2.DeleteRoute(&ec2.DeleteRouteInput{
-					RouteTableId:         aws.String(customRTB),
-					DestinationCidrBlock: aws.String("0.0.0.0/0"),
-				})
-			})
-		}
-
-		// Associate with the default subnet, then disassociate. The bash
-		// driver uses `default-for-az` to find a subnet; def.SubnetID
-		// is already that subnet (DiscoverDefaultVPC prefers DefaultForAz).
-		harness.Step(t, "associate-route-table %s <- %s", customRTB, def.SubnetID)
-		assocOut, err := c.EC2.AssociateRouteTable(&ec2.AssociateRouteTableInput{
-			RouteTableId: aws.String(customRTB),
-			SubnetId:     aws.String(def.SubnetID),
-		})
-		require.NoError(t, err, "associate-route-table custom")
-		assocID := aws.StringValue(assocOut.AssociationId)
-		require.NotEmpty(t, assocID, "custom AssociationId empty")
-		assocReleased := false
-		t.Cleanup(func() {
-			if !assocReleased {
-				_, _ = c.EC2.DisassociateRouteTable(&ec2.DisassociateRouteTableInput{
-					AssociationId: aws.String(assocID),
-				})
-			}
-		})
-		harness.Detail(t, "custom_assoc", assocID)
-
-		harness.Step(t, "disassociate-route-table %s", assocID)
-		_, err = c.EC2.DisassociateRouteTable(&ec2.DisassociateRouteTableInput{
-			AssociationId: aws.String(assocID),
-		})
-		require.NoError(t, err, "disassociate-route-table")
-		assocReleased = true
-
-		if routeAdded {
-			harness.Step(t, "delete-route 0.0.0.0/0")
-			_, err = c.EC2.DeleteRoute(&ec2.DeleteRouteInput{
-				RouteTableId:         aws.String(customRTB),
-				DestinationCidrBlock: aws.String("0.0.0.0/0"),
-			})
-			require.NoError(t, err, "delete-route 0.0.0.0/0")
-		}
-
-		harness.Step(t, "delete-route-table %s", customRTB)
-		_, err = c.EC2.DeleteRouteTable(&ec2.DeleteRouteTableInput{
-			RouteTableId: aws.String(customRTB),
-		})
-		require.NoError(t, err, "delete-route-table")
-		rtbDeleted = true
-	})
-
-	// --- Step 3: Error paths -----------------------------------------------
-	t.Run("Step3_ErrorPaths", func(t *testing.T) {
-		// Scratch RT for the "cannot delete local route" assertion. The bash
-		// version uses a fresh table so the failure doesn't poison the main
-		// table or the Step 2 lifecycle table.
-		harness.Step(t, "create-route-table (scratch for error paths)")
-		ctOut, err := c.EC2.CreateRouteTable(&ec2.CreateRouteTableInput{
-			VpcId: aws.String(def.VPCID),
-		})
-		require.NoError(t, err, "create scratch RT")
-		require.NotNil(t, ctOut.RouteTable, "create-route-table returned nil RouteTable")
-		errRTB := aws.StringValue(ctOut.RouteTable.RouteTableId)
-		require.NotEmpty(t, errRTB, "scratch RouteTableId empty")
-		t.Cleanup(func() {
-			_, _ = c.EC2.DeleteRouteTable(&ec2.DeleteRouteTableInput{
-				RouteTableId: aws.String(errRTB),
-			})
-		})
-
-		// Discover the scratch table's local CIDR (the VPC's CIDR block).
-		descScratch, err := c.EC2.DescribeRouteTables(&ec2.DescribeRouteTablesInput{
-			RouteTableIds: []*string{aws.String(errRTB)},
-		})
-		require.NoError(t, err, "describe scratch RT")
-		require.NotEmpty(t, descScratch.RouteTables, "scratch RT not visible")
-		scratchLocal := ""
-		for _, r := range descScratch.RouteTables[0].Routes {
-			if aws.StringValue(r.GatewayId) == "local" {
-				scratchLocal = aws.StringValue(r.DestinationCidrBlock)
-				break
-			}
-		}
-		require.NotEmptyf(t, scratchLocal,
-			"scratch RT %s missing local route", errRTB)
-
-		// DeleteRoute on the local CIDR must be rejected. Bash just asserts
-		// non-zero exit; we assert *any* error here rather than pin a code,
-		// because the gateway might surface this as InvalidParameterValue,
-		// OperationNotPermitted, or a custom "cannot delete local route" code.
-		// If the gateway settles on a single code in future we can tighten this
-		// to harness.AssertAWSError.
-		harness.Step(t, "delete-route local=%s (must fail)", scratchLocal)
-		_, err = c.EC2.DeleteRoute(&ec2.DeleteRouteInput{
-			RouteTableId:         aws.String(errRTB),
-			DestinationCidrBlock: aws.String(scratchLocal),
-		})
-		require.Errorf(t, err,
-			"delete-route on local route %s succeeded — must be rejected", scratchLocal)
-		// Sanity: the rejection should be an AWS-shaped error, not a transport
-		// error. Surface the code in the test log so a future tightening to
-		// AssertAWSError has a known target.
-		var aerr awserr.Error
-		if errors.As(err, &aerr) {
-			harness.Detail(t, "delete_local_err", aerr.Code())
-		}
-
-		// DeleteRouteTable on the main RT must be rejected — same shape.
-		harness.Step(t, "delete-route-table %s (main, must fail)", mainRTBID)
-		_, err = c.EC2.DeleteRouteTable(&ec2.DeleteRouteTableInput{
-			RouteTableId: aws.String(mainRTBID),
-		})
-		require.Errorf(t, err,
-			"delete-route-table on main RT %s succeeded — must be rejected", mainRTBID)
-		if errors.As(err, &aerr) {
-			harness.Detail(t, "delete_main_err", aerr.Code())
-		}
-	})
-}
-
-// runReplaceRouteConvergence validates ReplaceRoute's control-plane contract:
-// a route installed against one IGW target re-points to a different IGW, with
-// the route record converging to the new gateway. It then walks the negative
-// matrix (no-such-route, local-route, missing-target, unknown-IGW, cross-VPC
-// IGW) so each rejection path stays pinned to its AWS error code.
-//
-// V1 scope note: ReplaceRoute updates the route record but publishes no vpc
-// event, so it does NOT reprogram the OVN datapath. There is no datapath to
-// assert here — the contract under test is control-plane convergence only.
-// When ReplaceRoute begins emitting add/delete-igw-route events, extend this
-// with a guest-egress assertion mirroring runNATGateway.
-//
-// Owns two scratch VPCs + three IGWs end to end, so it never touches the
-// singleton or default VPC and is safe in the parallel bucket. No pool mode or
-// OVN needed — every step is pure KV control plane.
-func runReplaceRouteConvergence(t *testing.T, fix *Fixture) {
-	harness.Phase(t, "Single — ReplaceRoute Convergence (control plane)")
-
-	c := fix.AWS
-
-	// --- Primary VPC + two attached IGWs -----------------------------------
-	vpcID := createScratchVPC(t, c, "10.77.0.0/16")
-	igw1 := createAttachedIGW(t, c, vpcID)
-	igw2 := createAttachedIGW(t, c, vpcID)
-
-	// --- Secondary VPC + its own IGW (cross-VPC rejection target) ----------
-	otherVPCID := createScratchVPC(t, c, "10.88.0.0/16")
-	igwOther := createAttachedIGW(t, c, otherVPCID)
-
-	// --- Custom route table in the primary VPC -----------------------------
-	harness.Step(t, "create-route-table (vpc=%s)", vpcID)
-	// e2e:allow-create — scratch route table owned by this ReplaceRoute test.
-	rtOut, err := c.EC2.CreateRouteTable(&ec2.CreateRouteTableInput{VpcId: aws.String(vpcID)})
-	require.NoError(t, err, "create-route-table")
-	require.NotNil(t, rtOut.RouteTable, "create-route-table returned nil RouteTable")
-	rtbID := aws.StringValue(rtOut.RouteTable.RouteTableId)
-	require.NotEmpty(t, rtbID, "RouteTableId empty")
-	t.Cleanup(func() {
-		_, _ = c.EC2.DeleteRouteTable(&ec2.DeleteRouteTableInput{RouteTableId: aws.String(rtbID)})
-	})
-
-	const dest = "192.0.2.0/24"
-
-	// --- CreateRoute -> igw1 -----------------------------------------------
-	harness.Step(t, "create-route %s -> %s", dest, igw1)
-	_, err = c.EC2.CreateRoute(&ec2.CreateRouteInput{
-		RouteTableId:         aws.String(rtbID),
-		DestinationCidrBlock: aws.String(dest),
-		GatewayId:            aws.String(igw1),
-	})
-	require.NoError(t, err, "create-route %s -> igw1", dest)
-	require.Equalf(t, igw1, routeGateway(t, c, rtbID, dest),
-		"route %s should target igw1 after create", dest)
-
-	// --- ReplaceRoute -> igw2 (genuine target convergence) -----------------
-	harness.Step(t, "replace-route %s -> %s", dest, igw2)
-	_, err = c.EC2.ReplaceRoute(&ec2.ReplaceRouteInput{
-		RouteTableId:         aws.String(rtbID),
-		DestinationCidrBlock: aws.String(dest),
-		GatewayId:            aws.String(igw2),
-	})
-	require.NoError(t, err, "replace-route %s -> igw2", dest)
-	require.Equalf(t, igw2, routeGateway(t, c, rtbID, dest),
-		"route %s did not converge to igw2 after replace-route", dest)
-	harness.Detail(t, "convergence", dest+": "+igw1+" -> "+igw2)
-
-	// --- Negative matrix ---------------------------------------------------
-	// Unknown destination CIDR — route doesn't exist in the table.
-	harness.Step(t, "replace-route on absent route (must fail)")
-	harness.ExpectError(t, "InvalidRoute.NotFound", func() error {
-		_, e := c.EC2.ReplaceRoute(&ec2.ReplaceRouteInput{
-			RouteTableId:         aws.String(rtbID),
-			DestinationCidrBlock: aws.String("198.51.100.0/24"),
-			GatewayId:            aws.String(igw2),
-		})
-		return e
-	})
-
-	// Local route is immutable.
-	harness.Step(t, "replace-route on local route (must fail)")
-	harness.ExpectError(t, "InvalidParameterValue", func() error {
-		_, e := c.EC2.ReplaceRoute(&ec2.ReplaceRouteInput{
-			RouteTableId:         aws.String(rtbID),
-			DestinationCidrBlock: aws.String("10.77.0.0/16"),
-			GatewayId:            aws.String(igw2),
-		})
-		return e
-	})
-
-	// Missing target — V1 requires a GatewayId.
-	harness.Step(t, "replace-route with no target (must fail)")
-	harness.ExpectError(t, "InvalidParameterValue", func() error {
-		_, e := c.EC2.ReplaceRoute(&ec2.ReplaceRouteInput{
-			RouteTableId:         aws.String(rtbID),
-			DestinationCidrBlock: aws.String(dest),
-		})
-		return e
-	})
-
-	// Unknown IGW target.
-	harness.Step(t, "replace-route to non-existent IGW (must fail)")
-	harness.ExpectError(t, "InvalidInternetGatewayID.NotFound", func() error {
-		_, e := c.EC2.ReplaceRoute(&ec2.ReplaceRouteInput{
-			RouteTableId:         aws.String(rtbID),
-			DestinationCidrBlock: aws.String(dest),
-			GatewayId:            aws.String("igw-00000000000000000"),
-		})
-		return e
-	})
-
-	// IGW attached to a different VPC than the route table.
-	harness.Step(t, "replace-route to cross-VPC IGW %s (must fail)", igwOther)
-	harness.ExpectError(t, "InvalidParameterValue", func() error {
-		_, e := c.EC2.ReplaceRoute(&ec2.ReplaceRouteInput{
-			RouteTableId:         aws.String(rtbID),
-			DestinationCidrBlock: aws.String(dest),
-			GatewayId:            aws.String(igwOther),
-		})
-		return e
-	})
-
-	// Each rejection left the route untouched at igw2.
-	require.Equalf(t, igw2, routeGateway(t, c, rtbID, dest),
-		"route %s drifted off igw2 after the rejection matrix", dest)
-}
-
-// createScratchVPC creates a VPC with the given CIDR and registers its deletion.
-func createScratchVPC(t *testing.T, c *harness.AWSClient, cidr string) string {
-	t.Helper()
-	harness.Step(t, "create-vpc %s", cidr)
-	// e2e:allow-create — scratch VPC owned end to end by the caller.
-	out, err := c.EC2.CreateVpc(&ec2.CreateVpcInput{CidrBlock: aws.String(cidr)})
-	require.NoError(t, err, "create-vpc %s", cidr)
-	require.NotNil(t, out.Vpc, "create-vpc %s returned nil Vpc", cidr)
-	id := aws.StringValue(out.Vpc.VpcId)
-	require.NotEmpty(t, id, "VpcId empty for %s", cidr)
-	t.Cleanup(func() {
-		_, _ = c.EC2.DeleteVpc(&ec2.DeleteVpcInput{VpcId: aws.String(id)})
-	})
-	return id
-}
-
-// createAttachedIGW creates an internet gateway, attaches it to vpcID, and
-// registers detach+delete cleanup (LIFO-safe: detach runs before the owning
-// VPC's delete).
-func createAttachedIGW(t *testing.T, c *harness.AWSClient, vpcID string) string {
-	t.Helper()
-	// e2e:allow-create — scratch IGW owned by the caller's ReplaceRoute test.
-	out, err := c.EC2.CreateInternetGateway(&ec2.CreateInternetGatewayInput{})
-	require.NoError(t, err, "create-internet-gateway")
-	require.NotNil(t, out.InternetGateway, "create-internet-gateway returned nil IGW")
-	igwID := aws.StringValue(out.InternetGateway.InternetGatewayId)
-	require.NotEmpty(t, igwID, "InternetGatewayId empty")
-	_, err = c.EC2.AttachInternetGateway(&ec2.AttachInternetGatewayInput{
-		InternetGatewayId: aws.String(igwID),
-		VpcId:             aws.String(vpcID),
-	})
-	require.NoError(t, err, "attach-internet-gateway %s -> %s", igwID, vpcID)
-	t.Cleanup(func() {
-		_, _ = c.EC2.DetachInternetGateway(&ec2.DetachInternetGatewayInput{
-			InternetGatewayId: aws.String(igwID),
-			VpcId:             aws.String(vpcID),
-		})
-		_, _ = c.EC2.DeleteInternetGateway(&ec2.DeleteInternetGatewayInput{
-			InternetGatewayId: aws.String(igwID),
-		})
-	})
-	return igwID
-}
-
-// routeGateway returns the GatewayId of the route matching destCidr in rtbID,
-// failing the test if the route is absent.
-func routeGateway(t *testing.T, c *harness.AWSClient, rtbID, destCidr string) string {
-	t.Helper()
-	out, err := c.EC2.DescribeRouteTables(&ec2.DescribeRouteTablesInput{
-		RouteTableIds: []*string{aws.String(rtbID)},
-	})
-	require.NoError(t, err, "describe-route-tables %s", rtbID)
-	require.NotEmpty(t, out.RouteTables, "route table %s not found", rtbID)
-	for _, r := range out.RouteTables[0].Routes {
-		if aws.StringValue(r.DestinationCidrBlock) == destCidr {
-			return aws.StringValue(r.GatewayId)
-		}
-	}
-	t.Fatalf("route %s not found in table %s", destCidr, rtbID)
-	return ""
 }
